Add promu checksum command
This is essentially a reimplementation of sha256sum, but removing the
directory name, thus removing the need to use more involved cd/grep
usage in circle.yml files.

I'm not sure if it's worth to reimplement such functionality, but it
reduces the amount of shell scripts and it seems that what we're aiming
for with promu?