Feature: Added check_licenses Command to Promu
Problem
It's cumbersome to copy and paste the same shell script file into all of
the various prometheus repositories.  This makes it difficult to make
changes to one and update them everywhere.

Additionally the prometheus project has various different programming
languages such as java, go, ruby, bash, and html.  We need to be able to
run a license check against all of those source files so we need
a configurable way to include them in our check.

Solution
Put the license checking code into promu and use this utility to check
headers everywhere.