Do not discover user/host for reproducible builds
Allow to overwrite build user if SOURCE_DATE_EPOCH is set in order to
make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This PR was done while working on [reproducible builds for
openSUSE](https://en.opensuse.org/openSUSE:Reproducible_Builds).

Signed-off-by: Witek Bedyk <witold.bedyk@suse.com>